Issue #113 - Implement integration
Centralize the reconstitue_contour function to be a public
export from condition_messages.  This will allow it to be
used by the integration operations as well.
